requirements.txt
torch==2.11.0
torchvision==0.26.0
Pillow==12.1.1
PyYAML==6.0.3
pandas==2.3.3
matplotlib==3.10.8
